// src/components/repo/__generated__/Commits_history.graphql.ts
export type Commits_history$data = {
  readonly history: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly author: {
          readonly email: string | null;
          readonly name: string | null;
        } | null;
        readonly authoredDate: any;
        readonly committedDate: any;
        readonly message: string;
        readonly pushedDate: any | null;
        readonly url: any;
      } | null;
    } | null> | null;
    readonly pageInfo: {
      readonly endCursor: string | null;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null;
    };
    readonly totalCount: number;
  };
  readonly id: string;
  readonly " $fragmentType": "Commits_history";
};
export type Commits_history$key = {
  readonly " $data"?: Commits_history$data;
  readonly " $fragmentSpreads": FragmentRefs<"Commits_history">;
};